backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
supabase==2.23.3
pydantic-settings==2.6.0
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
requests==2.32.3

